Github comments fix
