+++ package.json
@@ -8,11 +8,15 @@
     "@testing-library/react": "^12.0.0",
     "@testing-library/user-event": "^13.2.1",
     "axios": "^0.21.1",
+    "bootstrap": "^5.1.3",
+    "formik": "^2.2.9",
     "react": "^17.0.2",
+    "react-bootstrap": "^2.1.1",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3"
+    "react-scripts": "^4.0.3",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-scripts start",
