+++ src/App.js
@@ -14,7 +14,7 @@ import ToysCampaign from './Campaigns/Toys/ToysCampaign';
 import MembersForm from './Components/Members/MembersForm';
 import ProjectsForm from './Components/Projects/ProjectsForm';
 import Organization from "./Components/Organization/Organization";
-import EditOrganization from './Components/Organization/EditOrganization'
+import OrganizationEdit from './Components/Organization/OrganizationEdit'
 
 function App() {
   return (
@@ -27,7 +27,7 @@ function App() {
           <Route path="/create-news" component={NewsForm} />
           <Route path="/backoffice/create-slide" component={SlidesForm} />
           <Route exact path="/backoffice/organization" component={Organization} />
-          <Route path="/backoffice/organization/edit" component={EditOrganization} />
+          <Route path="/backoffice/organization/edit" component={OrganizationEdit} />
           <Route path="/create-testimonials" component={TestimonialForm} />
           <Route path="/create-user" component={UserForm} />
           <Route path="/create-member" component={MembersForm} />
