+++ src/Helpers/formValidations.js
@@ -24,11 +24,4 @@ export const yupCustomString = (min, max, minMsg, maxMsg) =>
     .max(max, maxMsg)
     .required("Campo obligatorio");
 
-export const yupImages = () =>
-  Yup.mixed()
-    .required("Campo obligatorio")
-    .test(
-      "fileFormat",
-      "Extensión inválida. Solo archivos jpg o png",
-      (value) => ["image/jpeg", "image/png"].includes(value?.type)
-    );
+export const yupImages = () => Yup.string().ensure().required("Topic is required!");
