+++ src/Services/publicApiService.js
@@ -3,7 +3,7 @@ import { HOST } from './config';
 
 const config = {
     headers: {
-        Group: 1                //Aqui va el ID del equipo!!
+        Group: 130                //Aqui va el ID del equipo!!
     }
 }
 
@@ -13,7 +13,7 @@ const Get = () => {
     .catch(err => console.log(err))
 }
 
-export const getOrganization = async () => {
+export const getOrganizationData = async () => {
     try {
         const response = await axios.get(`${HOST}/organization`)
         return response.data
@@ -23,7 +23,7 @@ export const getOrganization = async () => {
     
 }
 
-export const putOrganization = async (body,id) => {
+export const updateputOrganizationData = async (body,id) => {
     try {
         const response = await axios.put(`${HOST}/organization/${id}`, body)
         return response.data
