+++ src/assets/index.js
@@ -1,0 +1,2 @@
+import logo from "./logo.png";
+export const LOGO = logo;
