--- src/assets/logos.js
@@ -1,2 +1,0 @@
-import logo1 from "./logo1.png";
-export const LOGOS = [{ logo: logo1 }];
