+++ src/rutas/config.js
@@ -1,0 +1,2 @@
+export const ORGANIZATION = "/backoffice/organization"
+export const EDIT_ORGANIZATION = "/backoffice/organization/edit"
